Initial commit
No history is shown beyond this commit because the project originally
had identifying information in it; this information has been scrubbed.
In addition, the source has been touched up to be more idiomatic Rust.
It was originally the submission for an assignment with a structure that
wasn't well-suited for a generic command-line app.